requirements.txt
appdirs==1.4.4
astroid==2.4.2
black==20.8b1
click==7.1.2
colorama==0.4.4
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.0
Pillow==8.0.1
pylint==2.6.0
regex==2020.10.23
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
wrapt==1.12.1
